Migrate beets logic into module/beets, leverage gql extension
