+++ app/gql/resolvers.ts
@@ -2,5 +2,13 @@ import { loadFilesSync } from '@graphql-tools/load-files';
 import path from 'path';
 import { mergeResolvers } from '@graphql-tools/merge';
 
-const resolversArray = loadFilesSync(path.join(__dirname, '../../modules/**/*.resolvers.*'));
-export const resolvers = mergeResolvers(resolversArray);
+const balancerResolversArray = loadFilesSync(path.join(__dirname, '../../modules/**/*.resolvers.*'), {
+    globOptions: { ignore: [path.join(__dirname, '../../modules/beets/*.resolvers.*')] },
+});
+
+const beetsResolversArray = loadFilesSync(path.join(__dirname, '../../modules/**/*.resolvers.*'), {
+    globOptions: { ignore: [path.join(__dirname, '../../modules/balancer/*.resolvers.*')] },
+});
+
+export const balancerResolvers = mergeResolvers(balancerResolversArray);
+export const beetsResolvers = mergeResolvers(beetsResolversArray);
