+++ app/loadRestRoutes.ts
@@ -1,5 +1,5 @@
 import { Express } from 'express';
-import { beetsGetCirculatingSupply } from '../modules/beets/lib/beets';
+import { beetsGetCirculatingSupply } from '../modules/beets/token/lib/beetsGetCirculatingSupply';
 import { tokenService } from '../modules/token/token.service';
 
 export function loadRestRoutes(app: Express) {
