+++ codegen.yml
@@ -117,13 +117,31 @@ generates:
         schema: ${USER_SNAPSHOT_SUBGRPAH}
         plugins:
             - schema-ast
-    graphql_schema_generated.ts:
-        schema: ./modules/**/*.gql
+    graphql_schema_generated_balancer.ts:
+        schema:
+            - ./modules/**/*.gql
+            - '!./modules/beets/beets.gql'
+            - '!./modules/**/*.beets.gql'
         plugins:
             - add:
                   content: |
                       import { gql } from 'apollo-server-express';
-                      export const schema = gql`
+                      export const balancerSchema = gql`
+                      #\n# THIS FILE IS AUTOGENERATED — DO NOT EDIT IT\n#
+            - schema-ast
+            - add:
+                  placement: 'append'
+                  content: '`;'
+    graphql_schema_generated_beets.ts:
+        schema:
+            - ./modules/**/*.gql
+            - '!./modules/balancer/balancer.gql'
+            - '!./modules/**/*.balancer.gql'
+        plugins:
+            - add:
+                  content: |
+                      import { gql } from 'apollo-server-express';
+                      export const beetsSchema = gql`
                       #\n# THIS FILE IS AUTOGENERATED — DO NOT EDIT IT\n#
             - schema-ast
             - add:
