+++ env.local
@@ -32,3 +32,6 @@ APOLLO_API_KEY=<not-required-locally>
 
 # Admin routes
 ADMIN_API_KEY=<chose-a-value>
+
+# The protocol to load
+PROTOCOL=beets | balancer
