+++ modules/beets/token/abi/FreshBeets.json
