+++ modules/pool/pool.resolvers.ts
@@ -66,23 +66,6 @@ const balancerResolvers: Resolvers = {
         poolGetLinearPools: async () => {
             return poolService.getGqlLinearPools();
         },
-        poolGetReliquaryFarmSnapshots: async (parent, { id, range }, context) => {
-            const snapshots = await poolService.getSnapshotsForReliquaryFarm(parseFloat(id), range);
-
-            return snapshots.map((snapshot) => ({
-                id: snapshot.id,
-                farmId: snapshot.farmId,
-                timestamp: snapshot.timestamp,
-                relicCount: `${snapshot.relicCount}`,
-                userCount: `${snapshot.userCount}`,
-                totalBalance: snapshot.totalBalance,
-                totalLiquidity: snapshot.totalLiquidity,
-                dailyDeposited: snapshot.dailyDeposited,
-                dailyWithdrawn: snapshot.dailyWithdrawn,
-                levelBalances: snapshot.levelBalances,
-                tokenBalances: snapshot.tokenBalances,
-            }));
-        },
     },
     Mutation: {
         poolSyncAllPoolsFromSubgraph: async (parent, {}, context) => {
@@ -247,13 +230,6 @@ const balancerResolvers: Resolvers = {
             isAdminRoute(context);
 
             await poolService.reloadAllTokenNestedPoolIds();
-
-            return 'success';
-        },
-        poolLoadReliquarySnapshotsForAllFarms: async (parent, args, context) => {
-            isAdminRoute(context);
-
-            await poolService.loadReliquarySnapshotsForAllFarms();
 
             return 'success';
         },
